refactor - replace graphkb/types.py and ipr/types.py with a single pori_python/types.py
